Dir structure changed & added seperate stylesheet
