+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text, TouchableOpacity } from 'react-native';
 
-import Signin from './screens/signin/Signin';
+import Signin from './screens/signin/index';
 
 function App() {
   return(
